Ship of Harkinian 64 bit compilation

+++ soh/src/code/z_play.c
@@ -198,8 +198,8 @@ void Gameplay_Init(GameState* thisx) {
     GraphicsContext* gfxCtx = globalCtx->state.gfxCtx;
     gGlobalCtx = globalCtx;
     //globalCtx->state.gfxCtx = NULL;
-    u32 zAlloc;
-    u32 zAllocAligned;
+    uintptr_t zAlloc;
+    uintptr_t zAllocAligned;
     size_t zAllocSize;
     Player* player;
     s32 playerStartCamId;
